requirements.txt
dominate==2.9.0
matplotlib==3.5.1
nni==3.0
numpy==1.21.5
pandas==1.4.2
Pillow==10.1.0
pretty_errors==1.2.25
Requests==2.31.0
scikit_image==0.19.2
scikit_learn==1.0.2
scipy==1.7.3
torch==2.1.1
torchsummary==1.5.1
torchvision==0.16.1
tqdm==4.64.0
visdom==0.2.4
